requirements.txt
flask
biopython

